+++ src/DTO/Requests/EditMessageTextRequestData.php
@@ -15,30 +15,33 @@ final readonly class EditMessageTextRequestData extends TelegramBotApiRequestDat
     /**
      * @param  array<string|int, mixed>|null  $entities
      * @param  TelegramBotData|array<string|int, mixed>|null  $linkPreviewOptions
+     * @param  TelegramBotData|array<string|int, mixed>|null  $richMessage
      * @param  TelegramBotData|array<string|int, mixed>|null  $replyMarkup
      * @param  array<string, mixed>  $extra
      */
     public static function make(
-        string $text,
         ?string $businessConnectionId = null,
         int|string|null $chatId = null,
         ?int $messageId = null,
         ?string $inlineMessageId = null,
+        ?string $text = null,
         string|TelegramParseMode|null $parseMode = null,
         ?array $entities = null,
         TelegramBotData|array|null $linkPreviewOptions = null,
+        TelegramBotData|array|null $richMessage = null,
         TelegramBotData|array|null $replyMarkup = null,
         array $extra = [],
     ): self {
         return new self(self::withoutNullValues(array_merge([
-            'text' => $text,
             'business_connection_id' => $businessConnectionId,
             'chat_id' => $chatId,
             'message_id' => $messageId,
             'inline_message_id' => $inlineMessageId,
+            'text' => $text,
             'parse_mode' => $parseMode,
             'entities' => $entities,
             'link_preview_options' => $linkPreviewOptions,
+            'rich_message' => $richMessage,
             'reply_markup' => $replyMarkup,
         ], $extra)));
     }